+++ src/quantum/azext_quantum/commands.py
@@ -2,45 +2,12 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
-import os
-
-from collections import OrderedDict
 
 # pylint: disable=line-too-long
+
+from collections import OrderedDict
 from azure.cli.core.commands import CliCommandType
-from azext_quantum._client_factory import cf_offerings
-
-from .operations.workspace import WorkspaceInfo
-from .operations.target import TargetInfo
-
-def validate_workspace_info(cmd, namespace):
-    group = getattr(namespace, 'resource_group_name', None)
-    name = getattr(namespace, 'workspace_name', None)
-    ws = WorkspaceInfo(cmd, group, name)
-
-    if not ws.subscription:
-        raise ValueError("Missing subscription argument")
-    if not ws.resource_group:
-        raise ValueError("Missing resource-group argument")
-    if not ws.name:
-        raise ValueError("Missing workspace name argument")
-
-def validate_target_info(cmd, namespace):
-    target_id = getattr(namespace, 'target_id', None)
-    target = TargetInfo(cmd, target_id)
-
-    if not target.target_id:
-        raise ValueError("Missing target-id argument")
-
-
-
-def validate_workspace_and_target_info(cmd, namespace):
-    validate_workspace_info(cmd, namespace)
-    validate_target_info(cmd, namespace)
-
-    if not 'AZURE_QUANTUM_STORAGE' in os.environ:
-        raise ValueError(f"Please set the AZURE_QUANTUM_STORAGE environment variable with an Azure Storage's connection string.")
-
+from ._validators import validate_workspace_info, validate_target_info, validate_workspace_and_target_info
 
 def transform_job(result):
     result = OrderedDict([
